Add FIPS services check in Provider put() method and debug info
The legacy “put()” method is for adding the services and it is used by
provider SUN and SunEC in Semeru OpenJDK8. From OpenJDK11 and above
versions, all those usages of “put()” method changed to "putService()"
method. But we still need to check the services which are added by the
legacy “put()” method in Java Security Restricted Mode. So, updated
the codes in Provider.java for checking the registered services, only
allowed the needed services be stored in the "legacyMap". And this
commit also added more debug info in Java Security Restricted Mode.

Signed-off-by: Tao Liu <tao.liu@ibm.com>