Remove unnecessary interrupt
The current implementation contains some code from the j9 threading
implementation. With openjdk threading we dont need a deadinterrupt
(which the old code is attempting to simulate) field as the interrupt
state is tracked in the JCL rather than the JVM.

Also, the old code actually clears the interrupt state which is
incorrect.

Signed-off-by: tajila <atobia@ca.ibm.com>