Get native EC key pointer during init through ECUtil instead of key impl
The acquisition of a native pointer for an EC key, private or public, is
being moved to the ECUtil, instead of the implementation classes of said
keys. This allows use of interop keys, rather than just keys coming from
SunEC.

The methods to get the pointers to the native keys are called during the
initialization phases, to discover the possibility of not being able to
utilize the native code and revert to the Java implementation early.

The engineUpdate() methods of the RawECDSA subclass are, also, updated
to use the java implementation, if one has been initialized.

Signed-off by: Kostas Tsiounis <kostas.tsiounis@ibm.com>