Fix thread interrupt status
It is possible for the Java and native interrupt status to get out of
sync. Sending two interrupts to the same thread in quick succession can
cause this. It is critical that clearing the interrupt status also
clears the native status. If the Java interrupt status is false, it is
important to verify that this is consistent with the native status to
prevent an unexpected interrupt exception the next time sleep() or
wait() is called.

Related to https://github.com/eclipse-openj9/openj9/issues/21804

Signed-off-by: tajila <atobia@ca.ibm.com>