Update Deterministic test to skip OpenJCEPlus
The OpenJCEPlus is added to the list of native
providers to be skipped when this test is run.

The test uses a custom SecureRandom instance
that is actually deterministic and expects the
same result, but this is not applicable in
OpenJCEPlus, since it uses a native library that
utilizes its own source of randomness.

The test is also removed from the various exclusion
lists.

Signed-off-by: Kostas Tsiounis <kostas.tsiounis@ibm.com>