Change bundled OpenSSL SONAME for Linux Platforms
JDK-bundled libcrypto-semeru.so and libssl-semeru.so used to
have the same SONAME as the operating system libcrypto.so and
libssl.so. When the dynamic linker tries to resolve a
libssl.so’s dependency on libcrypto.so.3, it will match the
already-loaded $JAVA_HOME/lib/libcrypto-semeru.so (whose SONAME
was also libcrypto.so.3). These two libcrypto.so.3 come from
different distributions and their symbol versions are
incompatible, this will hit an exception.

Signed-off-by: Jinhang Zhang <Jinhang.Zhang@ibm.com>